refactor: delete header-section and add SectionHeader component
